я нахуйййййййййййййййййййййй удаляю ломбок и больше никогда в жизни е…

+++ src/main/java/ru/noleg/crmmm/repository/GroupRepository.java
@@ -1,7 +1,7 @@
 package ru.noleg.crmmm.repository;
 
 import org.springframework.data.repository.CrudRepository;
-import ru.noleg.crmmm.repository.model.Group;
+import ru.noleg.crmmm.entity.Group;
 
-public interface GroupRepository extends CrudRepository<Group, Integer> {
+public interface GroupRepository extends CrudRepository<Group, Long> {
 }
